requirements.txt
tensorflow>=2.5.0
numpy>=1.19.5
matplotlib>=3.4.0
scikit-learn>=0.24.0
Pillow>=8.2.0
pandas>=1.3.0
seaborn>=0.11.1